requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
jinja2
gunicorn
python-multipart
sqlalchemy
websockets
watchdog
passlib[bcrypt]
bcrypt
python-jose[cryptography]
psycopg2-binary
python-dotenv
redis
broadcaster
google-generativeai
stripe
Pillow
